refactor: Mapping 클래스 분리
