+++ app/src/main/java/woowacourse/shopping/data/remote/source/CouponDataSourceImpl.kt
@@ -4,7 +4,7 @@ import woowacourse.shopping.data.remote.api.CouponApiService
 import woowacourse.shopping.data.remote.api.NetworkManager
 import woowacourse.shopping.data.source.CouponDataSource
 import woowacourse.shopping.domain.model.coupon.Coupon
-import woowacourse.shopping.utils.DtoMapper.toCoupon
+import woowacourse.shopping.utils.Mapper.CouponMapper.toCoupon
 
 class CouponDataSourceImpl(
     private val couponService: CouponApiService = NetworkManager.couponService(),
