+++ app/src/main/java/woowacourse/shopping/data/repository/RecentlyProductRepositoryImpl.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import woowacourse.shopping.data.db.recently.RecentlyProductDatabase
 import woowacourse.shopping.domain.model.product.RecentlyProduct
 import woowacourse.shopping.domain.repository.RecentlyProductRepository
-import woowacourse.shopping.utils.EntityMapper.toRecentlyProduct
-import woowacourse.shopping.utils.EntityMapper.toRecentlyProductEntity
+import woowacourse.shopping.utils.Mapper.ProductMapper.toRecentlyProduct
+import woowacourse.shopping.utils.Mapper.ProductMapper.toRecentlyProductEntity
 import woowacourse.shopping.utils.exception.ErrorEvent
 
 class RecentlyProductRepositoryImpl(context: Context) : RecentlyProductRepository {
